parrallel postproc, listlayout

+++ scripts/run_amfi_lsf.sh
@@ -4,11 +4,15 @@ queue="cccr-res"
 WCLOCK="240"
 JOBNAME="AMFI"
 STDOUT="stdout"_$JOBNAME
+
 submit_combine=True
+nproc_combine=3
+combine_only=True
+combine_child_run=0
+bypassrunning=True
 
-
-EXE=/moes/home/prajeesh/spec2d/exec/spec2d/spec2d.exe
-RUNNCCP2R=/moes/home/prajeesh/spec2d/exec/run_mppnccp2r/run_mppnccp2r
+EXE=_EXE_
+RUNNCCP2R=_ROOTDIR_/exec/run_mppnccp2r/run_mppnccp2r
 
 
 lsf=True
@@ -16,7 +20,7 @@ if ! [ -x "$(command -v bjobs)" ]; then
 	lsf=False
 fi
 
-
+if [ ! "$bypassrunning" == "True" ];then
 if [ "$lsf" == "True" ]; then
 	alljobs=$(bjobs -noheader -o 'exec_cwd jobid job_name' 2>/dev/null | grep "$(pwd) ")
 	if [ ! -z "$alljobs" ]; then
@@ -25,6 +29,7 @@ if [ "$lsf" == "True" ]; then
 		echo $alljobs
 		exit 1
 	fi
+fi
 fi
 
 line1=$(sed -n '/&atmos_nml/,/\//p' input.nml | \
@@ -46,9 +51,14 @@ done
 
 echo "NPES = "$npes 
 
+if [ "$npes" -eq "1" ]; then
+	submit_combine=False
+fi
+
 STDOUT=${STDOUT}
 
 tfile=$(mktemp)
+echo $tfile
 cat <<EOF > $tfile
 
 #!/bin/bash
@@ -73,21 +83,30 @@ EOF
 rm -f $STDOUT
 
 if [ "$submit_combine" == "True" ]; then
-	echo "Removing any previous unprocessed output files."
-	rm -f *.nc.????
+	if [ ! "$combine_only" == "True" ]; then
+		echo "Removing any previous unprocessed output files."
+		rm -f *.nc.????
+	fi
 fi
 
-if [ "$lsf" == "True" ]; then
-	output=$(bsub < $tfile)
-	echo $output
-	jobid=$(echo $output | head -n1 | cut -d'<' -f2 | cut -d'>' -f1;)
-	if [ "$jobid" -eq "$jobid" ] 2>/dev/null; then
-	  	echo "Job submitted" 
-	else
-	  	exit 1
-	fi
+if [ "$combine_only" == "True" ]; then
+	echo "Only combine"
+	COND=""
 else
-	bash $tfile
+  	cat $tfile
+	if [ "$lsf" == "True" ]; then
+		output=$(bsub < $tfile)
+		echo $output
+		jobid=$(echo $output | head -n1 | cut -d'<' -f2 | cut -d'>' -f1;)
+		if [ "$jobid" -eq "$jobid" ] 2>/dev/null; then
+		  	echo "Job submitted" 
+		else
+		  	exit 1
+		fi
+		COND="#BSUB -w started($jobid)"
+	else
+		bash $tfile
+	fi
 fi
 
 if [ "$submit_combine" == "True" ]; then
@@ -95,6 +114,7 @@ if [ "$submit_combine" == "True" ]; then
 WCLOCK=$((WCLOCK*2))
 
 tfile=$(mktemp)
+echo $tfile
 cat <<EOF > $tfile
 
 #!/bin/bash
@@ -104,18 +124,19 @@ cat <<EOF > $tfile
 #BSUB -W ${WCLOCK}:00
 #BSUB -J ${JOBNAME}_combine
 #BSUB -o ${STDOUT}_combine
-#BSUB -w started($jobid)
-#BSUB -n 1
+#$COND
+#BSUB -n $nproc_combine
 
 ulimit -c unlimited
 set -xu
 
-mpirun -n 1 $RUNNCCP2R <<< "&opts_nml removein=T, atmpes=$npes, child_run=T /"
-
+mpirun -n $nproc_combine $RUNNCCP2R \
+<<< "&opts_nml removein=1, atmpes=$npes, child_run=$combine_child_run /"
 EOF
 
 rm -f ${STDOUT}_combine
 
+cat $tfile
 if [ "$lsf" == "True" ]; then
 	bsub < $tfile
 else
